Добавил тесты.

+++ database/migrations/2025_07_21_202051_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->id();
             $table->string('name');
             $table->string('type');
-            $table->decimal('price', 10, 2);
+            $table->unsignedInteger('price');
             $table->string('description');
             $table->string('image');
             $table->timestamps();
